Split view classes into dedicated modules
