+++ app/views/index.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+
+class IndexView(TemplateView):
+    template_name = "index.html"
