+++ app/views/rikishi.py
@@ -1,23 +1,6 @@
-from django.views.generic import DetailView, ListView, TemplateView
+from django.views.generic import DetailView, ListView
 
-from .models import Division, Rikishi
-
-
-class IndexView(TemplateView):
-    template_name = "index.html"
-
-
-class DivisionListView(ListView):
-    model = Division
-    template_name = "division_list.html"
-
-
-class DivisionDetailView(DetailView):
-    model = Division
-    template_name = "division_detail.html"
-    slug_field = "name"
-    slug_url_kwarg = "slug"
-    context_object_name = "division"
+from ..models import Rikishi
 
 
 class RikishiListView(ListView):
